# Parser improved, about 35% faster. Fixes for nested array like output…