--- a/bench/bench.nim
+++ b/bench/bench.nim
@@ -1,0 +1,9 @@
+import benchy, bench_old, bench_new
+
+timeIt "old parser":
+  for i in 0 .. 1_000_000:
+    testOld()
+
+timeIt "new parser":
+  for i in 0 .. 1_000_000:
+    testNew()