--- a/redisparser.nimble
+++ b/redisparser.nimble
@@ -1,11 +1,11 @@
 # Package
 
-version       = "0.1.1"
+version       = "0.1.2"
 author        = "xmonader"
 description   = "RESP(REdis Serialization Protocol) Serialization for Nim"
 license       = "Apache2"
 srcDir        = "src"
-
+skipDirs      = @["tests", "bench"]
 # Dependencies
 
 requires "nim >= 0.18.0"