--- a/src/redisparser.nim
+++ b/src/redisparser.nim
@@ -1,7 +1,9 @@
-import strformat, tables, json, strutils, sequtils, hashes, net, asyncdispatch, asyncnet, os, strutils, parseutils, deques, options
+import strformat, strutils, hashes, net,  strutils
 
-const CRLF* = "\r\n"
-const REDISNIL = "\0\0"
+const
+  CRLF = "\r\n"
+  CRLF_LEN = len(CRLF)
+  REDIS_NIL = "\0\0"
 
 type
   ValueKind* = enum
@@ -15,7 +17,7 @@ type
     of vkBulkStr: bs*: string
     of vkArray: l*: seq[RedisValue]
 
-proc `$`*(obj: RedisValue): string = 
+proc `$`*(obj: RedisValue): string =
   case obj.kind
     of vkStr : return  fmt"{$(obj.s)}"
     of vkBulkStr: return fmt"{$(obj.bs)}"
@@ -23,7 +25,7 @@ proc `$`*(obj: RedisValue): string =
     of vkArray: return fmt"{$(obj.l)}"
     of vkError: return fmt"{$(obj.err)}"
 
-proc hash*(obj: RedisValue): Hash = 
+proc hash*(obj: RedisValue): Hash =
   case obj.kind
   of vkStr : !$(hash(obj.s))
   of vkBulkStr: !$(hash(obj.bs))
@@ -52,7 +54,7 @@ proc `==`* (a, b: RedisValue): bool =
       of vkError:
           result = a.err == b.err
 
-proc encode*(v: RedisValue) : string 
+proc encode*(v: RedisValue) : string
 proc encodeStr(v: RedisValue) : string =
   return fmt"+{v.s}{CRLF}"
 
@@ -65,16 +67,15 @@ proc encodeErr(v: RedisValue) : string =
 proc encodeInt(v: RedisValue) : string =
   return fmt":{v.i}{CRLF}"
 
-proc encodeArray(v: RedisValue): string = 
+proc encodeArray(v: RedisValue): string =
   var res = "*" & $len(v.l) & CRLF
   for el in v.l:
     res &= encode(el)
   res &= CRLF
   return res
 
-
 proc encode*(v: RedisValue) : string =
-  case v.kind 
+  case v.kind
   of vkStr: return encodeStr(v)
   of vkInt:    return encodeInt(v)
   of vkError:  return encodeErr(v)
@@ -82,115 +83,77 @@ proc encode*(v: RedisValue) : string =
   of vkArray: return encodeArray(v)
 
 
-proc decodeStr(s: string): (RedisValue, int) =
-  let crlfpos = s.find(CRLF)
-  return (RedisValue(kind:vkStr, s:s[1..crlfpos-1]), crlfpos+len(CRLF))
+proc decodeStr(s: string, pos: var int): RedisValue =
+  let crlfPos = s.find(CRLF, pos)
+  result = RedisValue(kind:vkStr, s: s[pos..<crlfPos])
+  pos = crlfPos + CRLF_LEN
 
-proc decodeBulkStr(s:string): (RedisValue, int) = 
-  let crlfpos = s.find(CRLF)
-  var bulklen = 0
-  let slen = s[1..crlfpos-1]
-  bulklen = parseInt(slen)
-  var bulk: string
-  if bulklen == -1:
-      bulk = REDISNIL 
-      return (RedisValue(kind:vkBulkStr, bs:bulk), crlfpos+len(CRLF))
-  # elif bulklen == 0:
-  #     bulk = ""
-  #     return (RedisValue(kind:vkBulkStr, bs:bulk), crlfpos+len(CRLF)+len(CRLF))
+proc decodeError(s: string, pos: var int): RedisValue =
+  let crlfPos = s.find(CRLF, pos)
+  result = RedisValue(kind:vkError, err: s[pos..<crlfPos])
+  pos = crlfPos + CRLF_LEN
+
+proc decodeBulkStr(s: string, pos: var int): RedisValue =
+  let
+    crlfPos = s.find(CRLF, pos)
+    bulkLen = parseInt(s[pos..<crlfPos])
+  pos = crlfPos + CRLF_LEN
+  result = RedisValue(kind:vkBulkStr)
+  if bulkLen == -1:
+    result.bs = REDIS_NIL
+  elif bulklen == 0:
+    inc(pos, CRLF_LEN)
   else:
-    let nextcrlf = s.find(CRLF, crlfpos+len(CRLF))
-    bulk = s[crlfpos+len(CRLF)..nextcrlf-1] 
-    return (RedisValue(kind:vkBulkStr, bs:bulk), nextcrlf+len(CRLF))
+    result.bs = newString(bulklen)
+    for i in 0..<bulklen:
+      result.bs[i] = s[pos + i]
+    inc(pos, bulklen + CRLF_LEN)
 
-proc decodeError(s: string): (RedisValue, int) =
-  let crlfpos = s.find(CRLF)
-  return (RedisValue(kind:vkError, err:s[1..crlfpos-1]), crlfpos+len(CRLF))
+proc decodeInt(s: string, pos: var int): RedisValue =
+  let
+    crlfPos = s.find(CRLF, pos)
+    ival = parseInt(s[pos..<crlfPos])
+  result = RedisValue(kind:vkInt, i: ival)
+  pos = crlfPos + CRLF_LEN
 
-proc decodeInt(s: string): (RedisValue, int) =
-  var i: int
-  let crlfpos = s.find(CRLF)
-  let sInt = s[1..crlfpos-1]
-  discard parseInt(sInt,i)
-  return (RedisValue(kind:vkInt, i:i), crlfpos+len(CRLF))
+proc decode(s: string, pos: var int): RedisValue
+proc decodeArray(s: string, pos: var int): RedisValue =
+  var
+    crlfPos = s.find(CRLF, pos)
+    arrLen = parseInt(s[pos..<crlfPos])
 
+  result = RedisValue(kind:vkArray)
+  if arrLen == -1:
+    pos = crlfPos + CRLF_LEN
+  else:
+    result.l = newSeq[RedisValue](arrLen)
+    pos = s.find(CRLF, crlfPos) + CRLF_LEN # next obj pos
 
-proc decode(s: string): (RedisValue, int)
+    var i = 0
+    while i < arrLen:
+      result.l[i] = decode(s, pos)
+      inc(i)
 
-proc decodeArray(s: string): (RedisValue, int) =
-  var arr = newSeq[RedisValue]()
-  var arrlen = 0
-  var crlfpos = s.find(CRLF)
-  var arrlenStr = s[1..crlfpos-1]
-  discard parseInt(arrlenStr,arrlen)
-  
-  var nextobjpos = s.find(CRLF)+len(CRLF)
-  var i = nextobjpos 
-  
-  if arrlen == -1:
-    
-    return (RedisValue(kind:vkArray, l:arr), i)
-  
-  while i < len(s) and len(arr) < arrlen:
-    var pair = decode(s[i..^1])
-    var obj = pair[0]
-    arr.add(obj)
-    i += pair[1]
-  return (RedisValue(kind:vkArray, l:arr), i+len(CRLF))
+proc decode(s: string, pos: var int): RedisValue =
+  let c = s[pos]
+  inc(pos)
+  case c
+  of '+':
+    return decodeStr(s, pos)
+  of '-':
+    return decodeError(s, pos)
+  of '$':
+    return decodeBulkStr(s, pos)
+  of ':':
+    return decodeInt(s, pos)
+  of '*':
+    return decodeArray(s, pos)
+  else:
+    raise newException(ValueError, fmt"Unreognized char {repr c}")
 
-
-proc decode(s: string): (RedisValue, int) =
-  var i = 0 
-  while i < len(s):
-    var curchar = $s[i]
-    var endindex = s.find(CRLF, i)+len(CRLF)
-    if endindex >= s.len:
-      endindex = s.len-1
-    if curchar == "+":
-      var pair = decodeStr(s[i..endindex])
-      var obj =  pair[0]
-      var count =  pair[1]
-      i += count
-      return (obj, i)
-    elif curchar == "-":
-      var pair = decodeError(s[i..endindex])
-      var obj =  pair[0]
-      var count =  pair[1]
-      i += count
-      return (obj, i)
-    elif curchar == "$":
-      var pair = decodeBulkStr(s[i..^1])
-      var obj =  pair[0]
-      var count =  pair[1]
-      i += count
-      return (obj, i)
-    elif curchar == ":":
-      var pair = decodeInt(s[i..endindex])
-      var obj =  pair[0]
-      var count =  pair[1]
-      i += count
-      return (obj, i)
-    elif curchar == "*":
-      var pair = decodeArray(s[i..^1])
-      let obj = pair[0]
-      let count =  pair[1]
-      i += count 
-      return (obj, i)
-    elif curchar == "\r":
-      i += 1
-      continue
-    elif curchar == "\n":
-      i += 1
-      continue
-    else:
-      echo fmt"Unreognized char {repr curchar}"
-      break
+const encodeValue* = encode
+proc decodeString*(resp: string): RedisValue =
+  var pos = 0
+  return decode(resp, pos)
 
 
-proc decodeResponse(resp: string): RedisValue = 
-  let pair = decode(resp)
-  return pair[0]
-
-let decodeForm = decodeResponse
-let encodeValue* = encode
-let decodeString* = decodeResponse