… of `XINFO STREAM`

Benchy result:
```
name ............................... min time      avg time    std dv   runs
old parser ...................... 3982.649 ms   3990.102 ms   ±10.541     x2
new parser ...................... 2584.027 ms   2590.411 ms    ±9.029     x2
```